+++ src/egon/data/airflow/dags/pipeline.py
@@ -3,7 +3,7 @@ from airflow.utils.dates import days_ago
 import airflow
 
 from egon.data.airflow.tasks import initdb
-from egon.data.data_import.openstreetmap import import_osm
+from egon.data.importing.openstreetmap import import_osm
 
 with airflow.DAG(
     "egon-data-processing-pipeline",
