+++ src/egon/data/importing/openstreetmap/oedb.style
